Add files via upload
Dimensionality Reduction Techniques:
PCA (Principal Component Analysis): PCA is applied to reduce the dimensionality of the feature space while preserving the most important information. It transforms the original features into a lower-dimensional space defined by the principal components.
SVD (Singular Value Decomposition): SVD is used to decompose a matrix into three separate matrices, enabling the reduction of feature space by selecting a subset of singular values and associated vectors.

Feature Selection: 
A feature selection technique is employed to identify and retain the most relevant features, discarding redundant or irrelevant ones. 

Models and Techniques:
Decision Tree Classifier: A decision tree classifier is trained on the preprocessed dataset to predict email categories. It partitions the feature space based on attribute values and classifies instances by traversing the tree structure.
Adaboost with Decision Tree as Base Classifier: Adaboost is used as an ensemble method with a decision tree as the base classifier. Adaboost iteratively trains multiple weak decision tree classifiers, focusing on misclassified instances from previous iterations.
Bagging with Decision Tree: Bagging is employed as an ensemble method with decision trees. Multiple decision trees are trained on bootstrap samples of the dataset, and the final prediction is obtained by aggregating the predictions from all the trees.
Random Forest: Random Forest is utilized as an ensemble method combining multiple decision trees. Each tree is trained on a random subset of features and samples from the dataset. The predictions of all trees are then aggregated to obtain the final result.
KNN with Voting Classifier: KNN (K-Nearest Neighbors) is used as a classification algorithm, and a Voting Classifier is employed to combine the predictions of multiple KNN models. The voting strategy can be based on majority voting or weighted voting.